修改主键泛型为Int类型

+++ src/main/java/com/lord/advertis/dao/AttachmentDao.java
@@ -3,6 +3,6 @@ package com.lord.advertis.dao;
 import com.lord.advertis.base.BaseMapper;
 import com.lord.advertis.domain.Attachment;
 
-public interface AttachmentDao extends BaseMapper<Attachment, Long>{
+public interface AttachmentDao extends BaseMapper<Attachment, Integer>{
 
 }
